+++ packages/db-postgres/src/migrateFresh.ts
@@ -79,6 +79,7 @@ export async function migrateFresh(
         err,
         msg: parseError(err, `Error running migration ${migration.name}. Rolling back`),
       })
+      process.exit(1)
     }
   }
 }