+++ packages/payload/src/collections/operations/delete.ts
@@ -104,12 +104,20 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
     // Retrieve documents
     // /////////////////////////////////////
 
-    const { docs } = await payload.db.find<GeneratedTypes['collections'][TSlug]>({
+    const dbArgs = {
       collection: collectionConfig.slug,
       locale,
       req,
       where: fullWhere,
-    })
+    }
+    let docs
+    if (collectionConfig?.db?.find) {
+      const result = await collectionConfig.db.find<GeneratedTypes['collections'][TSlug]>(dbArgs)
+      docs = result.docs
+    } else {
+      const result = await payload.db.find<GeneratedTypes['collections'][TSlug]>(dbArgs)
+      docs = result.docs
+    }
 
     const errors = []
 
@@ -160,7 +168,7 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
         // Delete document
         // /////////////////////////////////////
 
-        await payload.db.deleteOne({
+        const deleteOneArgs = {
           collection: collectionConfig.slug,
           req,
           where: {
@@ -168,7 +176,12 @@ async function deleteOperation<TSlug extends keyof GeneratedTypes['collections']
               equals: id,
             },
           },
-        })
+        }
+        if (collectionConfig?.db?.deleteOne) {
+          await collectionConfig.db.deleteOne(deleteOneArgs)
+        } else {
+          await payload.db.deleteOne(deleteOneArgs)
+        }
 
         // /////////////////////////////////////
         // afterRead - Fields