+++ packages/payload/src/collections/operations/deleteByID.ts
@@ -96,13 +96,19 @@ async function deleteByID<TSlug extends keyof GeneratedTypes['collections']>(
     // /////////////////////////////////////
     // Retrieve document
     // /////////////////////////////////////
-
-    const docToDelete = await req.payload.db.findOne({
+    let docToDelete: Document
+    const dbArgs = {
       collection: collectionConfig.slug,
       locale: req.locale,
       req,
       where: combineQueries({ id: { equals: id } }, accessResults),
-    })
+    }
+
+    if (collectionConfig?.db?.findOne) {
+      docToDelete = await collectionConfig.db.findOne(dbArgs)
+    } else {
+      docToDelete = await req.payload.db.findOne(dbArgs)
+    }
 
     if (!docToDelete && !hasWhereAccess) throw new NotFound(t)
     if (!docToDelete && hasWhereAccess) throw new Forbidden(t)
@@ -132,11 +138,17 @@ async function deleteByID<TSlug extends keyof GeneratedTypes['collections']>(
     // Delete document
     // /////////////////////////////////////
 
-    let result = await req.payload.db.deleteOne({
+    let result
+    const deleteOneArgs = {
       collection: collectionConfig.slug,
       req,
       where: { id: { equals: id } },
-    })
+    }
+    if (collectionConfig?.db?.deleteOne) {
+      result = await collectionConfig?.db.deleteOne(deleteOneArgs)
+    } else {
+      result = await payload.db.deleteOne(deleteOneArgs)
+    }
 
     // /////////////////////////////////////
     // Delete Preferences