+++ test/relationships/config.ts
@@ -281,6 +281,31 @@ export default buildConfigWithDefaults({
         },
       ],
     },
+    {
+      slug: 'pages',
+      fields: [
+        {
+          type: 'array',
+          name: 'menu',
+          fields: [
+            {
+              name: 'label',
+              type: 'text',
+            },
+          ],
+        },
+      ],
+    },
+    {
+      slug: 'rels-to-pages',
+      fields: [
+        {
+          name: 'page',
+          type: 'relationship',
+          relationTo: 'pages',
+        },
+      ],
+    },
   ],
   onInit: async (payload) => {
     await payload.create({